requirements.txt
biopython
cobra==0.17.1
pyparsing
pytest-cov
python-libsbml
